v1.2.17 Update
- Changed the data storage method from sending directly to database to
sending the data using HTTP to a server which then stores the data. This
allows for a much easier implementation of many different database and
storage types such as being able to use a PHP server. The new default is
to send data to a local HTTP server that runs as part of TrakHound
Server which then can respond to client requests much like a web server.
The data is backed up to a SQLite database where only the current day's
data is kept. A server application that stores historical data is
available as part of the TrakHound Standard package.
- The Overview dashboard page has been redesigned to use the dark
backgroud (upon customer requests) and can dynamically size itself based
on the window size and number of devices. This makes it look a lot
better when only a couple of devices are shown and also prevents the
user from having to scroll as much when many devices are configured.
This can be manually adjusted using the page zoom controls also.
- A new dashboard page called Status Grid has been added and is used to
view just the basic status (Active, Idle, Alert).

- A bug was fixed for the Device Manager's Edit Device page where the
pages weren't always being loaded correctly.
- The Add Device -> Auto Detect function no longer automatically starts
when the page is opened. Your IT department will thank us.
- The separate Client Enabled and Server Enabled have been combined into
a single Enabled option which can be set in the Device Manager page.
- A bug was fixed in the Configuration->AutoGenerate class to accept an
MTConnect Device that doesn't have any Description info.
- The Developer Console has been temporarily removed due to some issues
with logging using XML (the XML log files got corrupted too easily) so
logging is now done with regular text files for now.
- The variable names for Device Images have been changed as they didn't
make much sense before. This can lead to some user's previously set
images not loading and may have to be set again. Sorry.
- Device Manager now has a Backup feature that backs up the selected
devices to file. This replaces the previous Save function.